pyproject.toml
[project]
name = "shl-assessment-recommender"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.136.1",
    "huggingface-hub>=1.14.0",
    "langchain>=1.2.18",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.2.2",
    "langchain-groq>=1.1.2",
    "langchain-openai>=1.2.1",
    "langgraph>=1.1.10",
    "numpy>=2.4.4",
    "pinecone>=9.0.0",
    "pydantic>=2.13.4",
    "pydantic-settings>=2.14.1",
    "python-dotenv>=1.2.2",
    "rank-bm25>=0.2.2",
    "uvicorn[standard]>=0.46.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
